requirements.txt
langgraph
langchain
ipykernel
langchain-groq
python-dotenv
langchain-tavily
uvicorn
fastapi
python-multipart